% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pime.prevalence.R
\name{pime.prevalence}
\alias{pime.prevalence}
\title{Prevalences intervals}
\usage{
pime.prevalence(phylist)
}
\arguments{
\item{phylist}{Input, takes a list of phyloseq objects (one for each variable),
output from pime.split.by.variable()}
}
\description{
This function takes the output of pime.split.by.variable(). It estimates the highest prevalence
possible (no empty OTU table), calculates prevalence for taxa at 5% interval by increments of 5, until
maximum prevalence possible (no empty OTU table or droping samples).
After prevalence calculation, merges phyloseq objects for each interval and variable.
}
\examples{
phylist=pime.split.by.variable(restroom, "Environment")
prev=pime.prevalence(phylist)
}
\keyword{Prevalence}
