% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pime.split.by.variable.R
\name{pime.split.by.variable}
\alias{pime.split.by.variable}
\title{Per variable slices}
\usage{
pime.split.by.variable(physeq, variable, drop_zeroes = T)
}
\arguments{
\item{physeq}{The input file in phyloseq object format}

\item{variable}{Any variable present in the metadata to be analyzed "variable to split"}

\item{drop_zeroes}{Remove OTUs with zeros, default is TRUE. If FALSE, OTUs with zero counts will be kept.
Further analysis based on dissimilarity measurements (Bray-Curtis) might produce empty files.}
}
\description{
The used input file (in phyloseq object format) is sliced according to a metadata
variable present in the metadata file. Code extracted from metagMisc: v.0.0.4 - DOI 10.5281/zenodo.172500
}
\examples{
phylist=pime.split.by.variable(restroom, "Environment")
}
