% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pime.best.prevalence.R
\name{pime.best.prevalence}
\alias{pime.best.prevalence}
\title{Best Prevalence}
\usage{
pime.best.prevalence(prev.list, variable)
}
\arguments{
\item{prev.list}{List phyloseq objects with the calculated prevalences for each interval. The output of
pime.prevalence()}

\item{variable}{Any variable in the metadata to be analyzed}
}
\value{
OOB error
}
\description{
This function is the core of PIME. It uses ranger::ranger() to build random forests trees for sample classification
with each prevalence interval returned by pime.prevalence(). Model uses
}
\details{
The output is a table with OOB error rate,
the number of remaining OTUs and sequences after prevalence filtering.
}
\examples{
#Spliting data by "Environment"
phylist=pime.split.by.variable(restroom, "Environment")

#Computes prevalences for each treatment "Environment" separately
prev=pime.prevalence(phylist)

#Finds best prevalence interval
set.seed(2124)
pime.best.prevalence(prev, "Environment")

}
\keyword{OOB}
\keyword{classification}
\keyword{prevalence}
